Split ListSelector into three modules - selector, validator and transformer
